REstructured admin controls to make it easier to expand
