+++ pages/index.js
@@ -1,8 +1,8 @@
 import { useState } from "react";
 import LayoutAuthenticated from "../components/layout-authenticated";
 import MemberCard from "../components/member-card";
-import AdminControls from "../components/admin-controls";
 import { useSession } from "next-auth/react";
+import AdminControls from "../components/admin-controls/admin-controls";
 
 export default function Home({ initialMembers }) {
   const [members, setMembers] = useState(initialMembers);
@@ -12,11 +12,11 @@ export default function Home({ initialMembers }) {
     <LayoutAuthenticated>
       <div className="px-6 py-24 mx-auto max-w-[80%] sm:py-32 lg:px-8">
         <section className="mx-auto max-w-2xl sm:text-center">
-          <h1 className="text-3xl font-bold tracking-tight text-brand-dark sm:text-4xl dark:text-brand-light">
+          <h1 className="text-brand-dark dark:text-brand-light text-3xl font-bold tracking-tight sm:text-4xl">
             Meet other members
           </h1>
 
-          <p className="mt-6 text-lg leading-8 text-brand-dark-highlight dark:text-brand-light-highlight">
+          <p className="text-brand-dark-highlight dark:text-brand-light-highlight mt-6 text-lg leading-8">
             This is your space to find others in your community. Engage in
             conversations, share your ideas, and develop strong relationships.
             Step in, join the dialogue, and explore the community!
